FROMLIST: cpufreq: dt: Register an Energy Model
The Energy Model framework provides an API to register the active power
of CPUs. Call this API from the cpufreq-dt driver with an estimation
of the power as P = C * V^2 * f with C, V, and f respectively the
capacitance of the CPU and the voltage and frequency of the OPP.

The CPU capacitance is read from the "dynamic-power-coefficient" DT
binding (originally introduced for thermal/IPA), and the voltage and
frequency values from PM_OPP.

Change-Id: Id7b79ae3aafcc53574b850cb91a25240ebffbdd4
Cc: "Rafael J. Wysocki" <rjw@rjwysocki.net>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
(Removed "OPTIONAL" label from title. Applied from
https://lore.kernel.org/lkml/20180912091309.7551-15-quentin.perret@arm.com/)
Signed-off-by: Quentin Perret <quentin.perret@arm.com>
Signed-off-by: Lau <laststandrighthere@gmail.com>